+++ src/Hooks/ProcessResultInterface.php
@@ -1,6 +1,8 @@
 <?php
 namespace Consolidation\AnnotatedCommand\Hooks;
 
+use Consolidation\AnnotatedCommand\AnnotationData;
+
 /**
  * A result processor takes a result object, processes it, and
  * returns another result object.  For example, if a result object
@@ -21,5 +23,5 @@ interface ProcessResultInterface
      *
      * @return mixed $result
      */
-    public function process($result, array $args);
+    public function process($result, array $args, AnnotationData $annotationData);
 }