+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 ### 2.0.0 - 15 September 2016
 
 - Hooks with no command name now apply to all commands defined in the same class. This is a change of behavior from the 1.x branch, where hooks with no command name applied to a command with the same method name in a *different* class.
+- Update the interfaces ValidatorInterface, ProcessResultInterface and AlterResultInterface such that their methods are passed annotation data, just as the procedural callbacks are.
 
 
 ### 1.4.0 - 13 September 2016